Minor Bug Fixes

+++ database/migrations/2021_10_28_003617_create_materials_table.php
@@ -22,8 +22,8 @@ class CreateMaterialsTable extends Migration
             $table->string('uom');
             $table->string('mtyp');
             $table->string('crcy');
-            $table->integer('price');
-            $table->integer('per');
+            $table->bigInteger('price');
+            $table->bigInteger('per');
             $table->timestamps();
             $table->softDeletes();
         });
