+++ database/migrations/2021_10_29_192034_create_products_table.php
@@ -22,8 +22,8 @@ class CreateProductsTable extends Migration
             $table->string('uom');
             $table->string('mtyp');
             $table->string('crcy');
-            $table->integer('price');
-            $table->integer('per');
+            $table->bigInteger('price');
+            $table->bigInteger('per');
             $table->softDeletes();
             $table->timestamps();
         });
